requirements.txt
numpy
pandas
networkx
python-igraph
leidenalg
